backend/requirements.txt
annotated-types==0.7.0
anyio==4.5.2
bcrypt==4.1.3
black==24.8.0
boto3==1.34.76
botocore==1.34.76
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
dnspython==2.6.1
ecdsa==0.19.0
email-validator==2.1.1
fastapi==0.110.1
flake8==7.1.1
h11==0.14.0
idna==3.7
iniconfig==2.0.0
isort==5.13.2
jmespath==1.0.1
jq==1.10.0
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
motor==3.3.2
mypy==1.11.2
mypy_extensions==1.0.0
numpy==1.26.4
oauthlib==3.2.2
packaging==24.1
pandas==2.2.2
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.2.0
pluggy==1.5.0
pyasn1==0.6.1
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.8.0
pymongo==4.7.1
pytest==8.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.18
pytokens==0.3.0
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
s3transfer==0.10.1
s5cmd==0.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
typer==0.12.3
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
watchfiles==0.22.0